Add Funcionalidade
Novas Funcionalidades Referentes às permissões de usuário.

+++ resources/views/backend/blog/table.blade.php
@@ -27,27 +27,46 @@
                     {!! $post->publicationLabel() !!}
                 </td>
 
-                {{-- ADD permissoes em breve --}}
-                <td>
-                    <a href="{{ route('backend.blog.edit', $post->id) }}" class="btn btn-primary-yellow" title="Editar">
-                        <i class="align-middle" data-feather="edit"></i>
-                    </a>
-                </td>
-                <td>
-                    <a href="{{ route('backend.blog.show', $post->id) }}" class="btn btn-primary-yellow" title="Comentar">
-                        <i class="align-middle" data-feather="message-circle"></i>
-                    </a>
-                </td>
-                <td>
-                    {!! Form::open(['method' => 'DELETE', 'route' => ['backend.blog.destroy', $post->id]]) !!}
-                    @method('delete')
-                    {{ csrf_field() }}
-                    <button href="#" class="btn btn-dark text-yellow1" title="Mover para Reciclagem" type="submit"
-                        onclick="return confirm('Tem a certeza?')">
-                        <i class="align-middle" data-feather="trash"></i>
-                    </button>
-                    {!! Form::close() !!}
-                </td>
+                @if (check_user_permissions($request, 'Blog@edit', $post->id))
+                    <td>
+                        <a href="{{ route('backend.blog.edit', $post->id) }}" class="btn btn-primary-yellow"
+                            title="Editar">
+                            <i class="align-middle" data-feather="edit"></i>
+                        </a>
+                    </td>
+                @else
+                    <td>
+                        <a href="#" class="btn btn-primary-yellow disabled" title="Editar">
+                            <i class="align-middle" data-feather="edit"></i>
+                        </a>
+                    </td>
+
+                    {{-- COMENTAR --}}
+                    <td>
+                        <a href="{{ route('backend.blog.show', $post->id) }}" class="btn btn-primary-yellow"
+                            title="Comentar">
+                            <i class="align-middle" data-feather="message-circle"></i>
+                        </a>
+                    </td>
+                @endif
+
+
+                @if (check_user_permissions($request, 'Blog@destroy', $post->id))
+                    <td>
+                        {!! Form::open(['method' => 'DELETE', 'route' => ['backend.blog.destroy', $post->id]]) !!}
+                        @method('delete')
+                        {{ csrf_field() }}
+                        <button href="#" class="btn btn-dark text-yellow1" title="Mover para Reciclagem"
+                            type="submit" onclick="return confirm('Tem a certeza?')">
+                            <i class="align-middle" data-feather="trash"></i>
+                        </button>
+                        {!! Form::close() !!}
+                    </td>
+                @else
+                 <button type="button" onclick="return false" class="btn btn-dark text-yellow1">
+                    <i class="fa fa-trash"></i>
+                </button>
+                @endif
             </tr>
         @endforeach
